+++ packages/twenty-front/src/modules/settings/developers/components/SettingsDevelopersWebhookTableRow.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useTheme } from '@emotion/react';
 import styled from '@emotion/styled';
 
-import { WebhookFieldItem } from '@/settings/developers/types/WebhookFieldItem';
+import { WebhookFieldItem } from '@/settings/developers/types/webhook/WebhookFieldItem';
 import { IconChevronRight } from '@/ui/display/icon';
 import { SoonPill } from '@/ui/display/pill/components/SoonPill';
 import { TableCell } from '@/ui/layout/table/components/TableCell';
